Added post titles dynamically
We add blog post titles to the nav bar using JS. We could have
created a frontmatter in the JavaScript and then set a layout
to have liquid replace the tag for us. Instead, we just take
a function argument. Regardless here is the the former:

arbitrary.js:
---
layout: blank
---

blank.html:
---
---